chore(desktop): keep the probe out of the oversized plugin file
The automation probe is a test seam, not runtime behaviour, so it moves to
ProactiveAssistantsPlugin+SuggestionProbe.swift rather than growing an
already-ratcheted product file; the plugin returns to its exact baseline.

DesktopAutomationBridge.swift does grow by the 13-line action registration —
registrations must live in that file's registry — so its baseline is raised with
a justification rather than silently.

Also replaces force-unwraps in the new date tests with XCTUnwrap (swiftlint), and
covers the new file in the screen-recording-permission e2e flow.

make preflight: 19/19 checks pass.
