+++ .github/scripts/product_file_line_count_ratchet_baseline/desktop-swift-root.json
@@ -2,7 +2,7 @@
   "files": {
     "desktop/macos/Desktop/Sources/AuthService.swift": 2844,
     "desktop/macos/Desktop/Sources/CloudConnectorFormAutomation.swift": 1678,
-    "desktop/macos/Desktop/Sources/DesktopAutomationBridge.swift": 4547,
+    "desktop/macos/Desktop/Sources/DesktopAutomationBridge.swift": 4560,
     "desktop/macos/Desktop/Sources/DesktopDiagnosticsManager.swift": 1516,
     "desktop/macos/Desktop/Sources/MemoryExportService.swift": 1528,
     "desktop/macos/Desktop/Sources/OmiApp.swift": 1590
@@ -10,7 +10,7 @@
   "raise_justifications": {
     "desktop/macos/Desktop/Sources/AuthService.swift": "Apple first-auth name capture signals observers (authNameDidUpdate) and persists the name to Firebase so it survives reinstalls. +32 keeps the durable accepted-deletion cleanup journal and owner-bound VM/sync drain inside the fenced sign-out commit so cleanup resumes after restart and credentials or owner visibility cannot change first (SCA-280).",
     "desktop/macos/Desktop/Sources/CloudConnectorFormAutomation.swift": "Connector form automation remains in the owning desktop integration boundary so explicit user actions can complete the provider flow without duplicating browser/session state.",
-    "desktop/macos/Desktop/Sources/DesktopAutomationBridge.swift": "debug_hit_probe diagnostic action (topmost window + hit-tested view at a screen point) registers beside the other non-prod debug actions; it found the notch click-sink dead zone.",
+    "desktop/macos/Desktop/Sources/DesktopAutomationBridge.swift": "probe_suggestion_nudge registration: the suggestion delivery path cannot otherwise be exercised without holding a leisure window frontmost for 30s, which takes the user's focus; action registrations must live in this file's registry.",
     "desktop/macos/Desktop/Sources/DesktopDiagnosticsManager.swift": "Adds the bounded shared-capture silent-microphone terminal incident so a persistent zero-sample CoreAudio route is queryable without recording sensitive content. +1 registers account-cutover retained-control fallback telemetry in the existing bounded diagnostic area.",
     "desktop/macos/Desktop/Sources/MemoryExportService.swift": "Separates MCP data serving from production OAuth/grant authority so Beta can use dev data endpoints without issuing or reading customer grants through the dev identity plane. Pinned swift-format normalizes line layout without changing runtime behavior, and cloudOAuthGrantClientIDs keep ChatGPT directory grants on the production client for dev/Beta builds.",
     "desktop/macos/Desktop/Sources/OmiApp.swift": "UserDefaults.didChangeNotification observer must document and implement the queue:nil + explicit-main-hop contract (synchronous main-queue delivery deadlocked the signed-out launch, #11396)."
